+++ anyblok_pyramid_rest_api/adapter.py
@@ -22,6 +22,8 @@ class Adapter:
         self.filters = {}
         self.orders_by = {}
         self.tags = {}
+        self.grouped_tags = {}
+        self.group_by_tags = {}
 
     def load_decorators(self):
         for attr, value in self.__class__.__dict__.items():
@@ -32,6 +34,10 @@ class Adapter:
                 self.orders_by[value.is_order_by] = attr
             elif hasattr(value, 'is_tag'):
                 self.tags[value.is_tag] = attr
+            elif hasattr(value, 'is_grouped_tag'):
+                self.grouped_tags[value.is_grouped_tag['name']] = attr
+                for tag in value.is_grouped_tag['tags']:
+                    self.group_by_tags[tag] = value.is_grouped_tag['name']
 
     def has_filter_for(self, key, operator):
         return True if self.filters.get(key, {}).get(operator) else False
@@ -48,8 +54,14 @@ class Adapter:
     def has_tag_for(self, tag):
         return True if self.tags.get(tag) else False
 
+    def has_grouped_tag_for(self, tag):
+        return True if self.group_by_tags.get(tag) else False
+
     def get_tag_for(self, tag):
         return getattr(self, self.tags[tag])
+
+    def get_grouped_tag_for(self, group):
+        return getattr(self, self.grouped_tags[group])
 
     @classmethod
     def filter(cls, key, operators):
@@ -73,7 +85,18 @@ class Adapter:
     @classmethod
     def tag(cls, name):
         def wrapper(method):
-            method.is_tag = name
+            if hasattr(method, 'is_grouped_tag'):
+                method.is_grouped_tag['tags'].append(name)
+            else:
+                method.is_tag = name
+            return method
+
+        return wrapper
+
+    @classmethod
+    def grouped_tag(cls, name):
+        def wrapper(method):
+            method.is_grouped_tag = {'name': name, 'tags': []}
             return method
 
         return wrapper