+++ anyblok_pyramid_rest_api/test_bloks/test_4/views.py
@@ -26,9 +26,11 @@ class CustomerAdapter(Adapter):
         return query
 
     @Adapter.tag('green')
-    def tag_is_green(self, querystring, query):
+    @Adapter.tag('blue')
+    @Adapter.grouped_tag('color')
+    def tag_is_green_or_blue(self, querystring, query, tags):
         query = query.join(self.registry.Customer.tags, aliased=True)
-        query = query.filter(self.registry.Tag.name == 'green')
+        query = query.filter(self.registry.Tag.name.in_(tags))
         return query
 
     @Adapter.tag('orange')